requirements.txt
astroid==2.0.4
certifi==2018.10.15
chardet==3.0.4
colorama==0.4.0
dj-database-url==0.5.0
dj-static==0.0.6
Django==2.1.5
djangorestframework==3.9.0
future==0.17.1
gunicorn==19.9.0
idna==2.8
isort==4.3.4
lazy-object-proxy==1.3.1
line-bot-sdk==1.8.0
mccabe==0.6.1
numpy==1.15.4
psycopg2==2.7.6.1
Pygments==2.3.1
pymongo==3.7.2
pytz==2018.7
requests==2.21.0
scikit-learn==0.20.2
scipy==1.1.0
six==1.11.0
static3==0.7.0
urllib3==1.24.1
virtualenv==16.0.0
wincertstore==0.2
wrapt==1.10.11
mongoengine==0.16.3
django-rest-framework-mongoengine==3.3.1